+++ lib/core/use_cases/remote/extras/get_tv_show_genres_case.dart
@@ -1,0 +1,11 @@
+import 'package:screen_pal/core/entities/extras/genre.dart';
+import 'package:screen_pal/core/services/remote/extras_repo.dart';
+
+final class GetTvShowGenresCase {
+  const GetTvShowGenresCase({required ExtrasRepo extrasRepo})
+      : _extrasRepo = extrasRepo;
+
+  final ExtrasRepo _extrasRepo;
+
+  Future<List<Genre>> execute() => _extrasRepo.getTvShowGenres();
+}