+++ lib/interfaces/providers/remote/extras/genres_providers.dart
@@ -4,17 +4,17 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 import 'package:screen_pal/infrastructures/container/container.dart' as service;
 import 'package:screen_pal/core/entities/extras/genre.dart';
-import 'package:screen_pal/core/use_cases/remote/extras/get_movie_genres.dart';
-import 'package:screen_pal/core/use_cases/remote/extras/get_tv_show_genres.dart';
+import 'package:screen_pal/core/use_cases/remote/extras/get_movie_genres_case.dart';
+import 'package:screen_pal/core/use_cases/remote/extras/get_tv_show_genres_case.dart';
 
 part 'genres_providers.g.dart';
 
 @Riverpod(keepAlive: true, dependencies: [])
 Future<List<Genre>> movieGenres(MovieGenresRef ref) {
-  return service.locator<GetMovieGenres>().execute();
+  return service.locator<GetMovieGenresCase>().execute();
 }
 
 @Riverpod(keepAlive: true, dependencies: [])
 Future<List<Genre>> tvShowGenres(TvShowGenresRef ref) {
-  return service.locator<GetTvShowGenres>().execute();
+  return service.locator<GetTvShowGenresCase>().execute();
 }