+++ lib/interfaces/providers/remote/movies/movie_list_providers.dart
@@ -10,22 +10,22 @@ part 'movie_list_providers.g.dart';
 
 @Riverpod(keepAlive: true)
 Future<List<Movie>> nowPlayingMovies(NowPlayingMoviesRef ref) async {
-  return service.locator<GetNowPlayingMovies>().execute();
+  return service.locator<GetNowPlayingMoviesCase>().execute();
 }
 
 @Riverpod(keepAlive: true)
 Future<List<Movie>> popularMovies(PopularMoviesRef ref) {
-  return service.locator<GetPopularMovies>().execute();
+  return service.locator<GetPopularMoviesCase>().execute();
 }
 
 @Riverpod(keepAlive: true)
 Future<List<Movie>> topRatedMovies(TopRatedMoviesRef ref) {
-  return service.locator<GetTopRatedMovies>().execute();
+  return service.locator<GetTopRatedMoviesCase>().execute();
 }
 
 @Riverpod(keepAlive: true)
 Future<List<Movie>> upcomingMovies(UpcomingMoviesRef ref) {
-  return service.locator<GetUpcomingMovies>().execute();
+  return service.locator<GetUpcomingMoviesCase>().execute();
 }
 
 @riverpod
@@ -33,5 +33,5 @@ Future<List<Movie>> movieRecommendations(
   MovieRecommendationsRef ref,
   int movieId,
 ) {
-  return service.locator<GetMovieRecommendations>().execute(movieId);
+  return service.locator<GetMovieRecommendationsCase>().execute(movieId);
 }