+++ lib/interfaces/router/router.dart
@@ -8,18 +8,20 @@ import 'package:screen_pal/interfaces/widgets/errors/app_http_error_widget.dart'
 
 final router = GoRouter(
   initialLocation: '/movies',
-  routerNeglect: true,
-  errorBuilder: (context, router) => Scaffold(
-    body: AppHttpErrorWidget(
-      statusCode: 404,
-      message: 'No resource found on "${router.uri.path}" path.',
-      action: ElevatedButton.icon(
-        onPressed: () => context.go('/'),
-        icon: const Icon(Icons.home),
-        label: const Text('Back to Home'),
+  routes: $appRoutes,
+  errorBuilder: (context, router) {
+    return Scaffold(
+      body: AppHttpErrorWidget(
+        statusCode: 404,
+        message: 'No resource found on "${router.uri.path}" path.',
+        action: ElevatedButton.icon(
+          onPressed: () => const MoviesRoute().go(context),
+          icon: const Icon(Icons.home),
+          label: const Text('Back to Home'),
+        ),
       ),
-    ),
-  ),
+    );
+  },
   redirect: (_, router) {
     final currentRoute = router.uri.path;
 
@@ -27,5 +29,4 @@ final router = GoRouter(
 
     return null;
   },
-  routes: routes,
 );