+++ lib/interfaces/widgets/others/products/products_carousel.dart
@@ -7,7 +7,7 @@ import 'package:screen_pal/common/constants.dart';
 import 'package:screen_pal/core/entities/movies/movie.dart';
 import 'package:screen_pal/core/entities/products/product.dart';
 import 'package:screen_pal/interfaces/providers/remote/extras/extras_providers.dart';
-import 'package:screen_pal/interfaces/router/app_navigator.dart';
+import 'package:screen_pal/interfaces/router/routes.dart';
 import 'package:screen_pal/interfaces/widgets/common/common_network_image.dart';
 import 'package:screen_pal/interfaces/utils/m3_text_styles.dart';
 
@@ -146,8 +146,8 @@ class _ProductsCarouselState extends ConsumerState<ProductsCarousel> {
 
 VoidCallback _onItemTap(BuildContext context, Product product) {
   return () => (product is Movie)
-      ? AppNavigator.movieDetail(context, product.id)
-      : AppNavigator.tvShowDetail(context, product.id);
+      ? MovieDetailRoute(movieId: product.id).go(context)
+      : TvShowDetailRoute(tvShowId: product.id).go(context);
 }
 
 class _ExtrasText extends ConsumerWidget {