+++ CHANGELOG.md
@@ -10,14 +10,17 @@
   - core
     - re-organize folders
     - redefine [AppSettingsRepo] into [ThemeModeCache] instead.
-    - Rename all "use case" classes with "Case" suffix (eg. [GetThemeModeCase] instead of [GetThemeMode]).
+      #b437da2ea18ab60d12f1b8ef82aa60f7b11b4eae
+    - Rename all "use case" classes with "Case" suffix (eg. [GetThemeModeCase]
+      instead of [GetThemeMode]). #807c72c505b17f1e564ec183d9107b006b79ef36
   - infrastrucutures
+    - re-organize folders. #86acabd933fad083eae31fffbc42718be62a8a5e
     - rename "service_locator" into "container".
   - interfaces
     - re-organize folders.
     - apply infrastructures changes.
     - Remove [AppNavigator], use typed routes to navigate instead.
-      #0911282e1c75e34b40bb994eed23e6c7aeccef06
+      #dcf00daba45ef5cd2e2b2375c0ae89a15192bc64
 
 ## v1.0.1